pyproject.toml
[project]
name = "lora-stuff"
version = "0.1.0"
description = "LoRA fine-tuning experiments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.20.0",
    "datasets>=2.14.0",
    "evaluate>=0.4.0",
    "huggingface-hub>=0.16.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "peft>=0.5.0",
    "pyarrow>=12.0.0",
    "safetensors>=0.3.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "sentencepiece>=0.1.99",
    "tensorboard>=2.14.0",
    "tokenizers>=0.13.0",
    "torch>=2.0.0",
    "tqdm>=4.65.0",
    "transformers>=4.30.0",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
include = ["ftdatasets*", "experiments*", "app*"]

